+++ tests/test_elasticsearch_api.py
@@ -2,9 +2,11 @@
 
 import json
 
-from ska_sdp_dataproduct_api.core.helperfunctions import DPDAPIStatus
-from ska_sdp_dataproduct_api.core.settings import METADATA_ES_SCHEMA_FILE
-from ska_sdp_dataproduct_api.elasticsearch.elasticsearch_api import ElasticsearchMetadataStore
+from ska_sdp_dataproduct_api.components.elasticsearch.elasticsearch_api import (
+    ElasticsearchMetadataStore,
+)
+from ska_sdp_dataproduct_api.configuration.settings import METADATA_ES_SCHEMA_FILE
+from ska_sdp_dataproduct_api.utilities.helperfunctions import DPDAPIStatus
 from tests.mock_elasticsearch_api import MockElasticsearch
 
 DPD_API_Status = DPDAPIStatus()
@@ -12,7 +14,7 @@ DPD_API_Status = DPDAPIStatus()
 
 def test_create_schema():
     """Method to test creation of schema."""
-    metadata_store = ElasticsearchMetadataStore(DPD_API_Status)
+    metadata_store = ElasticsearchMetadataStore()
     metadata_store.es_client = MockElasticsearch()
 
     with open(
@@ -28,7 +30,7 @@ def test_create_schema():
 
 def test_insert_metadata():
     """Method to test insertion of metadata."""
-    metadata_store = ElasticsearchMetadataStore(DPD_API_Status)
+    metadata_store = ElasticsearchMetadataStore()
     metadata_store.es_client = MockElasticsearch()
 
     with open(
@@ -46,7 +48,7 @@ def test_insert_metadata():
 
 def test_update_dataproduct_list():
     """Method to test insertion of metadata."""
-    metadata_store = ElasticsearchMetadataStore(DPD_API_Status)
+    metadata_store = ElasticsearchMetadataStore()
     metadata_store.es_client = MockElasticsearch()
 
     with open(
@@ -73,7 +75,7 @@ def test_update_dataproduct_list():
 
 def test_search_metadata():
     """Method to test search of metadata"""
-    metadata_store = ElasticsearchMetadataStore(DPD_API_Status)
+    metadata_store = ElasticsearchMetadataStore()
     metadata_store.es_client = MockElasticsearch()
     metadata_store.es_client.ping = lambda: True
     metadata_list = metadata_store.search_metadata(
@@ -97,7 +99,7 @@ def test_search_metadata():
 
 def test_search_metadata_default_value():
     """Method to test search of metadata if metadata_key_value_pair is None"""
-    metadata_store = ElasticsearchMetadataStore(DPD_API_Status)
+    metadata_store = ElasticsearchMetadataStore()
     metadata_store.es_client = MockElasticsearch()
     metadata_store.es_client.ping = lambda: True
     metadata_list = metadata_store.search_metadata(
@@ -121,7 +123,7 @@ def test_search_metadata_default_value():
 
 def test_search_metadata_blank_list():
     """Method to test search of metadata if blank list is given."""
-    metadata_store = ElasticsearchMetadataStore(DPD_API_Status)
+    metadata_store = ElasticsearchMetadataStore()
     metadata_store.es_client = MockElasticsearch()
     metadata_store.es_client.ping = lambda: True
     metadata_list = metadata_store.search_metadata(
@@ -146,7 +148,7 @@ def test_search_metadata_blank_list():
 def test_search_metadata_no_value():
     """Method to test search of metadata
     if metadata_key_value_pair is not given"""
-    metadata_store = ElasticsearchMetadataStore(DPD_API_Status)
+    metadata_store = ElasticsearchMetadataStore()
     metadata_store.es_client = MockElasticsearch()
     metadata_store.es_client.ping = lambda: True
     metadata_list = metadata_store.search_metadata(start_date="2020-01-01", end_date="2100-01-01")